requirements.txt
pathspec==0.12.1
pyperclip==1.9.0
PyYAML==6.0.2
